+++ test-resources/with-cyclic-dependency-namespaced/src/sample/controller/foo.clj
@@ -1,2 +1,0 @@
-(ns sample.controller.foo
-  (:require [sample.logic.foo :as logic.foo]))