+++ test-resources/with-violations-namespaced/src/sample/controller/foo.clj
@@ -1,1 +1,0 @@
-(ns sample.controller.foo)