+++ test-resources/without-violations-namespaced/src/sample/logic/bar.clj
@@ -1,2 +1,0 @@
-(ns without-violations.src.sample.logic.bar
-  (:require [sample.logic.foo :as logic.foo]))