+++ test/clj_depend/analyzer_test.clj
@@ -24,12 +24,12 @@
                       :c {:defined-by         ".*\\.c\\..*"
                           :accessed-by-layers #{:a :b}}}})
 
-(def config-with-namespaced-layers {:layers {:a {:namespaces         #{'foo.a.bar}
-                                                 :accessed-by-layers #{}}
-                                             :b {:namespaces         #{'foo.b.bar 'foo.b.baz}
-                                                 :accessed-by-layers #{:a}}
-                                             :c {:defined-by         ".*\\.c\\..*"
-                                                 :accessed-by-layers #{:a :b}}}})
+(def config-with-namespaces {:layers {:a {:namespaces         #{'foo.a.bar}
+                                          :accessed-by-layers #{}}
+                                      :b {:namespaces         #{'foo.b.bar 'foo.b.baz}
+                                          :accessed-by-layers #{:a}}
+                                      :c {:defined-by         ".*\\.c\\..*"
+                                          :accessed-by-layers #{:a :b}}}})
 
 (def ns-deps (concat ns-deps-a ns-deps-b ns-deps-c))
 (def namespaces (map :name ns-deps))
@@ -68,11 +68,11 @@
                               :namespaces       namespaces-with-violations
                               :dependency-graph dependency-graph-with-violations})))))
 
-(deftest analyze-namespaced-test
+(deftest analyze-config-with-namespaces-test
 
   (testing "should return zero violations when there is no forbidden access"
     (is (= []
-           (analyzer/analyze {:config           config-with-namespaced-layers
+           (analyzer/analyze {:config           config-with-namespaces
                               :namespaces       namespaces
                               :dependency-graph dependency-graph}))))
 
@@ -80,6 +80,6 @@
     (is (= [{:namespace 'foo.c.bar
              :dependency-namespace 'foo.b.bar
              :message "\"foo.c.bar\" should not depends on \"foo.b.bar\""}]
-           (analyzer/analyze {:config           config-with-namespaced-layers
+           (analyzer/analyze {:config           config-with-namespaces
                               :namespaces       namespaces-with-violations
                               :dependency-graph dependency-graph-with-violations})))))