+++ test/clj_depend/api_integration_test.clj
@@ -1,66 +1,71 @@
 (ns clj-depend.api-integration-test
   (:require [clj-depend.api :as api]
+            [clj-depend.config :as config]
             [clojure.java.io :as io]
             [clojure.test :refer :all]))
 
-(deftest analyze
-
+(deftest analyze-invalid-options
   (testing "should thrown an exception when project-root is not a file"
     (is (thrown? AssertionError
                  (api/analyze {:project-root "./sample"}))))
 
   (testing "should thrown an exception when project-root is not an existent file"
     (is (thrown? AssertionError
-                 (api/analyze {:project-root (io/file "./non-existent-folder")}))))
+                 (api/analyze {:project-root (io/file "./non-existent-folder")})))))
 
+(deftest analyze-project-with-cyclic-dependency
+  (testing "should fail when a cyclic dependency is identified"
+    (is (= {:result-code 2
+            :message     "Circular dependency between sample.logic.foo and sample.controller.foo"}
+           (api/analyze {:project-root (io/file (io/resource "with-cyclic-dependency"))})))))
+
+(deftest analyze-project-without-violations
   (testing "should succeed when there are no violations"
     (is (= {:result-code 0
             :message     "No violations found!"}
-           (api/analyze {:project-root (io/file (io/resource "without-violations"))}))))
+           (api/analyze {:project-root (io/file (io/resource "without-violations"))})))))
 
+(deftest analyze-project-with-violation
   (testing "should fail when there is at least one violation"
     (is (= {:result-code 1
             :message     "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""
             :violations  [{:namespace            'sample.logic.foo
                            :dependency-namespace 'sample.controller.foo
                            :message              "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
-           (api/analyze {:project-root (io/file (io/resource "with-violations"))}))))
+           (api/analyze {:project-root (io/file (io/resource "with-violations"))})))))
 
+(deftest analyze-project-with-custom-config
   (testing "should fail when given a different configuration as a parameter"
     (is (= {:result-code 1
-            :message     "\"sample.controller.foo\" should not depends on \"sample.logic.foo\""
-            :violations  [{:namespace            'sample.controller.foo
-                           :dependency-namespace 'sample.logic.foo
-                           :message              "\"sample.controller.foo\" should not depends on \"sample.logic.foo\""}]}
+            :message     "\"sample.server\" should not depends on \"sample.controller.foo\""
+            :violations  [{:namespace            'sample.server
+                           :dependency-namespace 'sample.controller.foo
+                           :message              "\"sample.server\" should not depends on \"sample.controller.foo\""}]}
            (api/analyze {:project-root (io/file (io/resource "without-violations"))
-                         :config       {:source-paths #{"src"}
-                                        :layers       {:controller {:defined-by         ".*\\.controller\\..*"
-                                                                    :accessed-by-layers #{}}
-                                                       :logic      {:defined-by         ".*\\.logic\\..*"
-                                                                    :accessed-by-layers #{}}}}}))))
+                         :config       (-> (config/read! (io/file (io/resource "without-violations")))
+                                           (assoc-in [:layers :controller :accessed-by-layers] #{}))}))))
 
   (testing "should succeed when given a different configuration as a parameter"
     (is (= {:result-code 0
             :message     "No violations found!"}
            (api/analyze {:project-root (io/file (io/resource "with-violations"))
                          :config       {:source-paths #{"src"}
-                                        :layers       {:controller {:defined-by         ".*\\.controller\\..*"
-                                                                    :accessed-by-layers #{:logic}}
-                                                       :logic      {:defined-by         ".*\\.logic\\..*"
-                                                                    :accessed-by-layers #{}}}}}))))
+                                        :layers       (-> (config/read! (io/file (io/resource "with-violations")))
+                                                          (assoc-in [:layers :controller :accessed-by-layers] #{:logic})
+                                                          (assoc-in [:layers :logic :accessed-by-layers] #{}))}}))))
 
   (testing "should fail when given a different configuration as a parameter without source-paths"
     (is (= {:result-code 1
-            :message     "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""
-            :violations  [{:namespace            'sample.logic.foo
+            :message     "\"sample.server\" should not depends on \"sample.controller.foo\""
+            :violations  [{:namespace            'sample.server
                            :dependency-namespace 'sample.controller.foo
-                           :message              "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
-           (api/analyze {:project-root (io/file (io/resource "with-violations"))
-                         :config       {:layers {:controller {:defined-by         ".*\\.controller\\..*"
-                                                              :accessed-by-layers #{}}
-                                                 :logic      {:defined-by         ".*\\.logic\\..*"
-                                                              :accessed-by-layers #{:controller}}}}}))))
+                           :message              "\"sample.server\" should not depends on \"sample.controller.foo\""}]}
+           (api/analyze {:project-root (io/file (io/resource "without-violations"))
+                         :config       (-> (config/read! (io/file (io/resource "without-violations")))
+                                           (dissoc :source-paths)
+                                           (assoc-in [:layers :controller :accessed-by-layers] #{}))})))))
 
+(deftest analyze-project-with-specific-namespace
   (testing "should succeed when the namespace that has the violation is not included in the analysis"
     (is (= {:result-code 0
             :message     "No violations found!"}
@@ -74,8 +79,9 @@
                            :dependency-namespace 'sample.controller.foo
                            :message              "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
            (api/analyze {:project-root (io/file (io/resource "with-violations"))
-                         :namespaces   #{'sample.logic.foo}}))))
+                         :namespaces   #{'sample.logic.foo}})))))
 
+(deftest analyze-project-with-specific-file
   (testing "should succeed when the files that has the violation is not included in the analysis"
     (is (= {:result-code 0
             :message     "No violations found!"}
@@ -89,98 +95,12 @@
                            :dependency-namespace 'sample.controller.foo
                            :message "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
            (api/analyze {:project-root (io/file (io/resource "with-violations"))
-                         :files        #{(io/file (io/resource "with-violations/src/sample/logic/foo.clj"))}}))))
+                         :files        #{(io/file (io/resource "with-violations/src/sample/logic/foo.clj"))}})))))
 
+(deftest analyze-project-with-specific-file-and-namespace
   (testing "should succeed when the files that has the violation is included but namespace that has the violation is not included in the analysis"
     (is (= {:result-code 0
             :message     "No violations found!"}
            (api/analyze {:project-root (io/file (io/resource "with-violations"))
                          :files        #{(io/file (io/resource "with-violations/src/sample/logic/foo.clj"))}
-                         :namespaces   #{'sample.controller.foo}}))))
-
-  (testing "should fail when a cyclic dependency is identified"
-    (is (= {:result-code 2
-            :message     "Circular dependency between sample.logic.foo and sample.controller.foo"}
-           (api/analyze {:project-root (io/file (io/resource "with-cyclic-dependency"))
-                         :source-paths #{"src"}})))))
-
-(deftest analyze-namespaced
-
-  (testing "should succeed when there are no violations"
-    (is (= {:result-code 0
-            :message     "No violations found!"}
-           (api/analyze {:project-root (io/file (io/resource "without-violations-namespaced"))}))))
-
-  (testing "should fail when there is at least one violation"
-    (is (= {:result-code 1
-            :message     "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""
-            :violations  [{:namespace            'sample.logic.foo
-                           :dependency-namespace 'sample.controller.foo
-                           :message              "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))}))))
-
-  (testing "should fail when given a different configuration as a parameter"
-    (is (= {:result-code 1
-            :message     "\"sample.controller.foo\" should not depends on \"sample.logic.foo\""
-            :violations  [{:namespace            'sample.controller.foo
-                           :dependency-namespace 'sample.logic.foo
-                           :message              "\"sample.controller.foo\" should not depends on \"sample.logic.foo\""}]}
-           (api/analyze {:project-root (io/file (io/resource "without-violations-namespaced"))
-                         :config       {:source-paths #{"src"}
-                                        :layers       {:controller {:namespaces         #{'sample.controller.foo}
-                                                                    :accessed-by-layers #{}}
-                                                       :logic      {:defined-by         ".*\\.logic\\..*"
-                                                                    :accessed-by-layers #{}}}}}))))
-
-  (testing "should succeed when given a different configuration as a parameter"
-    (is (= {:result-code 0
-            :message     "No violations found!"}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))
-                         :config       {:source-paths #{"src"}
-                                        :layers       {:controller {:namespaces         #{'sample.controller.foo}
-                                                                    :accessed-by-layers #{:logic}}
-                                                       :logic      {:defined-by         ".*\\.logic\\..*"
-                                                                    :accessed-by-layers #{}}}}}))))
-
-  (testing "should succeed when the namespace that has the violation is not included in the analysis"
-    (is (= {:result-code 0
-            :message     "No violations found!"}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))
-                         :namespaces   #{'sample.controller.foo}}))))
-
-  (testing "should fail even when only the namespace that has the violation is included in the analysis."
-    (is (= {:result-code 1
-            :message     "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""
-            :violations  [{:namespace            'sample.logic.foo
-                           :dependency-namespace 'sample.controller.foo
-                           :message              "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))
-                         :namespaces   #{'sample.logic.foo}}))))
-
-  (testing "should succeed when the files that has the violation is not included in the analysis"
-    (is (= {:result-code 0
-            :message     "No violations found!"}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))
-                         :files        #{(io/file (io/resource "with-violations-namespaced/src/sample/controller/foo.clj"))}}))))
-
-  (testing "should fail even when only the files that has the violation is included in the analysis."
-    (is (= {:result-code 1
-            :message     "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""
-            :violations  [{:namespace 'sample.logic.foo
-                           :dependency-namespace 'sample.controller.foo
-                           :message "\"sample.logic.foo\" should not depends on \"sample.controller.foo\""}]}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))
-                         :files        #{(io/file (io/resource "with-violations-namespaced/src/sample/logic/foo.clj"))}}))))
-
-  (testing "should succeed when the files that has the violation is included but namespace that has the violation is not included in the analysis"
-    (is (= {:result-code 0
-            :message     "No violations found!"}
-           (api/analyze {:project-root (io/file (io/resource "with-violations-namespaced"))
-                         :files        #{(io/file (io/resource "with-violations-namespaced/src/sample/logic/foo.clj"))}
-                         :namespaces   #{'sample.controller.foo}}))))
-
-  (testing "should fail when a cyclic dependency is identified"
-    (is (= {:result-code 2
-            :message     "Circular dependency between sample.logic.foo and sample.controller.foo"}
-           (api/analyze {:project-root (io/file (io/resource "with-cyclic-dependency-namespaced"))
-                         :source-paths #{"src"}})))))
+                         :namespaces   #{'sample.controller.foo}})))))